requirements.txt
anthropic
pdfplumber
faiss-cpu
numpy
python-dotenv
sentence-transformers